make project structure more consistent
