+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import Head from 'next/head';
 import { GetServerSideProps } from 'next';
 import { Article, User } from '@/types';
-import { getArticles } from '@/resolvers/articles';
+import { getArticles } from '@/services/articles';
 import { ArticlePreview } from '@/components/ArticlePreview';
 import { Header } from '@/components/Header';
 import { Footer } from '@/components/Footer';
